Create archive lmd-ghost impl for simplicity
